+++ components/common/index.ts
@@ -8,14 +8,10 @@ export { default as Logo } from './Logo';
 export { default as HeadMeta } from './HeadMeta';
 
 export { default as QuizCard } from './QuizCard';
-export { default as QuizList } from './QuizList';
-export { default as QuizItem } from './QuizItem';
 export { default as MyQuizCard } from './MyQuizCard';
 export { default as RecentQuiz } from './RecentQuiz';
 export { default as PopularQuiz } from '../popular/PopularQuiz';
 export { default as SkeletonQuizCard } from './SkeletonQuizCard';
-
-export { default as RankingBoard } from './RankingBoard';
 
 export { default as SideNavigationBar } from './SideNavigationBar';
 export { default as Icon } from './Icon';
