+++ components/partials/main/AllQuizSection.tsx
