+++ components/partials/solve/main/ChoiceImage.tsx
