+++ components/partials/solve/main/index.ts
@@ -1,0 +1,4 @@
+export { default as QuizList } from './QuizList';
+export { default as ChoiceImage } from './ChoiceImage';
+export { default as ChoiceText } from './ChoiceText';
+export { default as QuizItem } from './QuizItem';
