+++ components/partials/solve/result/UserScore.tsx
