+++ components/ranking/RankCard.tsx
