+++ components/ranking/index.ts
@@ -1,0 +1,3 @@
+export { default as RankCard } from './RankCard';
+export { default as Ranker } from './Ranker';
+export { default as RankingBoard } from './RankingBoard';
